pylint and rename caikit_example/caikit_library --> caikit_huggingface_demo/runtime
* By convention match repo name and source dir (dashes become underscores)
* Replace confusing "caikit_library" with "runtime".
* image-classification go back to input image (not url) consistent with the other computer vision tasks.
* Make it pass pylint, black, and isort.

Some pylint checks were disabled to reduce the amount of changes and will be reintroduced later (e.g. line length).
