--- caikit_example/models/sentiment/config.yml
@@ -1,1 +1,0 @@
-block_id: FADC770C-25C8-4685-A176-51FF67B382C1
