+++ caikit_huggingface_demo/client/object_detection.py
@@ -45,7 +45,7 @@ class ObjectDetection:
             metadata=[("mm-model-id", model)],
         )
 
-        results = [x for x in response.objects]
+        results = list(response.objects)
         results.sort(
             key=lambda c: c.score, reverse=True
         )  # Sort so numbering will be in desc score order
