+++ caikit_huggingface_demo/example_models_extras/embeddings/config.yml
