+++ caikit_huggingface_demo/example_models_extras/image_classification/config.yml
