+++ caikit_huggingface_demo/example_models_extras/image_segmentation/config.yml
