+++ caikit_huggingface_demo/example_models_extras/object_detection/config.yml
