+++ caikit_huggingface_demo/runtime/data_model/image_segmentation.py
