+++ caikit_huggingface_demo/runtime/data_model/object_detection.py
