+++ caikit_huggingface_demo/runtime/tasks/conversational.py
@@ -13,13 +13,13 @@
 # limitations under the License
 
 # Third Party
-from caikit_library.block_ids import CONVERSATIONAL
-from caikit_library.data_model.results import Text
-from caikit_library.hf_base import HFBase
+from block_ids import CONVERSATIONAL
+from runtime.data_model.results import Text
+from runtime.hf_base import HFBase
 from transformers import Conversation, pipeline
 
 # Local
-from caikit.core import ModuleLoader, block
+from caikit.core import block
 
 TASK = "conversational"
 
@@ -33,13 +33,12 @@ DEFAULT_MODEL_REVISION = "4e936e3a11f8e077b31eec8f045499c92c7cf087"
 class Conversational(HFBase):
     def __init__(self, model_config_path) -> None:
         super().__init__()
-        loader = ModuleLoader(model_config_path)
         hf_model, hf_revision = self.read_config(
             model_config_path, DEFAULT_MODEL, DEFAULT_MODEL_REVISION
         )
         self.pipe = pipeline(task=TASK, model=hf_model, revision=hf_revision)
 
-    def run(self, text_in: str) -> Text:
+    def run(self, text_in: str) -> Text:  # pylint: disable=arguments-differ
         conversation = Conversation()
         conversation.add_user_input(text_in)
         conversation = self.pipe(conversation)
