+++ caikit_huggingface_demo/runtime/tasks/image_classification.py
@@ -12,19 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License
 
-# Standard
-import os
-
 # Third Party
-from caikit_library.block_ids import IMAGE_CLASSIFICATION
-from caikit_library.data_model.classification import ClassificationPrediction, ClassInfo
-from caikit_library.hf_base import HFBase
-from PIL import Image
+from block_ids import IMAGE_CLASSIFICATION
+from runtime.data_model.classification import ClassificationPrediction, ClassInfo
+from runtime.hf_base import HFBase
 from transformers import pipeline
-import requests
 
 # Local
-from caikit.core import ModuleLoader, ModuleSaver, block
+from caikit.core import block
 
 TASK = "image-classification"
 # DEFAULTS: google/vit-base-patch16-224 and revision 5dca96d
@@ -37,25 +32,21 @@ class ImageClassification(HFBase):
 
     def __init__(self, model_config_path) -> None:
         super().__init__()
-        loader = ModuleLoader(model_config_path)
         hf_model, hf_revision = self.read_config(model_config_path, None, None)
         self.pipe = pipeline(task=TASK, model=hf_model, revision=hf_revision)
 
-    def run(self, url_in: str) -> ClassificationPrediction:
+    def run(
+        self, encoded_bytes_or_url: str
+    ) -> ClassificationPrediction:  # pylint: disable=arguments-differ
         """Run HF sentiment analysis
         Args:
-            url_in Image in bytes
+            encoded_bytes_or_url Encoded image bytes (or url string)
         Returns:
             ClassificationPrediction: predicted classes with their confidence score.
         """
 
-        if os.path.isfile(url_in):
-            image = Image.open(url_in)
-        else:
-            image = Image.open(requests.get(url_in, stream=True).raw)
-
+        image = HFBase.get_image_bytes(encoded_bytes_or_url)
         raw_results = self.pipe(image)  # , top_k=9)
-        print("RAW RESULTS: ", raw_results)
         class_info = []
         for result in raw_results:
             class_info.append(
