+++ tox.ini
@@ -17,9 +17,9 @@ deps =
     -rrequirements.txt
 commands =
     {3.8,3.11}-pytest: python -c "print('No tests to run yet')"
-    # pytest {tty:--color=yes} {posargs:tests} --cov=caikit_example --cov-report=html
-    {3.8,3.11}-pylint: python -m pylint caikit_example
-    {3.8,3.11}-black-check: python -m black --check --diff caikit_example
-    {3.8,3.11}-black-fix: python -m black caikit_example
-    {3.8,3.11}-isort-check: python -m isort --check-only --diff caikit_example
-    {3.8,3.11}-isort-fix: python -m isort --atomic caikit_example
+    # pytest {tty:--color=yes} {posargs:tests} --cov=caikit_huggingface_demo --cov-report=html
+    {3.8,3.11}-pylint: python -m pylint caikit_huggingface_demo
+    {3.8,3.11}-black-check: python -m black --check --diff caikit_huggingface_demo
+    {3.8,3.11}-black-fix: python -m black caikit_huggingface_demo
+    {3.8,3.11}-isort-check: python -m isort --check-only --diff caikit_huggingface_demo
+    {3.8,3.11}-isort-fix: python -m isort --atomic caikit_huggingface_demo
